requirements.txt
asgiref==3.8.1
Django==4.2
django-bootstrap5==23.4
django-multiselectfield==0.1.12
sqlparse==0.4.4
